// src/api/generated/model/note.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * acid
 * API for acidarchive.com
 * OpenAPI spec version: 0.0.1-alpha.8
 */

export enum Note {
  C= 'C',
  'C#'= 'C#',
  D= 'D',
  'D#'= 'D#',
  E= 'E',
  F= 'F',
  'F#'= 'F#',
  G= 'G',
  'G#'= 'G#',
  A= 'A',
  'A#'= 'A#',
  B= 'B',
  Chigh= 'Chigh',

}