// src/api/generated/model/tB303Pattern.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * acid
 * API for acidarchive.com
 * OpenAPI spec version: 0.0.1-alpha.4
 */
import type { TB303PatternAccent } from './tB303PatternAccent';
import type { TB303PatternAuthor } from './tB303PatternAuthor';
import type { TB303PatternCreatedAt } from './tB303PatternCreatedAt';
import type { TB303PatternCutOffFreq } from './tB303PatternCutOffFreq';
import type { TB303PatternDecay } from './tB303PatternDecay';
import type { TB303PatternDescription } from './tB303PatternDescription';
import type { TB303PatternEnvMod } from './tB303PatternEnvMod';
import type { TB303PatternId } from './tB303PatternId';
import type { TB303PatternIsPublic } from './tB303PatternIsPublic';
import type { TB303PatternResonance } from './tB303PatternResonance';
import type { TB303Step } from './tB303Step';
import type { TB303PatternTempo } from './tB303PatternTempo';
import type { TB303PatternTitle } from './tB303PatternTitle';
import type { TB303PatternTriplets } from './tB303PatternTriplets';
import type { TB303PatternTuning } from './tB303PatternTuning';
import type { TB303PatternUpdatedAt } from './tB303PatternUpdatedAt';
import type { TB303PatternWaveform } from './tB303PatternWaveform';

export interface TB303Pattern {
  accent?: TB303PatternAccent;
  author?: TB303PatternAuthor;
  created_at?: TB303PatternCreatedAt;
  cut_off_freq?: TB303PatternCutOffFreq;
  decay?: TB303PatternDecay;
  description?: TB303PatternDescription;
  env_mod?: TB303PatternEnvMod;
  id?: TB303PatternId;
  is_public?: TB303PatternIsPublic;
  name: string;
  resonance?: TB303PatternResonance;
  steps: TB303Step[];
  tempo?: TB303PatternTempo;
  title?: TB303PatternTitle;
  triplets?: TB303PatternTriplets;
  tuning?: TB303PatternTuning;
  updated_at?: TB303PatternUpdatedAt;
  waveform?: TB303PatternWaveform;
}